Shared tests for bitvectors
