+++ src/rl_vector/tests.rs
@@ -1,0 +1,41 @@
+use super::*;
+
+//-----------------------------------------------------------------------------
+
+// FIXME random runs, random bitvector, random rl_vector, empty vector, full vector
+
+//-----------------------------------------------------------------------------
+
+// FIXME construction: full runs, partial runs, bit-by-bit
+
+//-----------------------------------------------------------------------------
+
+// FIXME try_iter, try_run_iter
+
+// FIXME basic: empty, non-empty, conversions, uniform, access, iter, serialize, large
+
+//-----------------------------------------------------------------------------
+
+// FIXME try_rank
+
+// FIXME rank: empty, non-empty, uniform, large
+
+//-----------------------------------------------------------------------------
+
+// FIXME try_select, try_one_iter
+
+// FIXME select: empty, non-empty, uniform, large
+
+//-----------------------------------------------------------------------------
+
+// FIXME try_select_zero, try_zero_iter
+
+// FIXME select_zero: empty, non-empty, uniform, large
+
+//-----------------------------------------------------------------------------
+
+// FIXME try_pred_succ
+
+// FIXME pred_succ: empty, non-empty, uniform, large
+
+//-----------------------------------------------------------------------------
